feat: delete Favorite and Watched movies routes

+++ db/movies.js
@@ -6,11 +6,11 @@ const findFavoritesByUserID = async (id, skip, limit) =>
 const findWatchedByUserID = async (id, skip, limit) =>
   await WatchedMovies.find({ userID: id }).skip(skip).limit(limit);
 
-const findFavoriteByTitle = async (title) =>
-  await FavoriteMovies.findOne({ title: title });
+const findFavoriteByMovieID = async (userID, movieID) =>
+  await FavoriteMovies.findOne({ userId: userID, idbId: movieID });
 
-const findWatchedByTitle = async (title) =>
-  await WatchedMovies.findOne({ title: title });
+const findWatchedByMovieID = async (userID, movieID) =>
+  await WatchedMovies.findOne({ userId: userID, idbId: movieID });
 
 const addFavoriteMovie = async (movie, id) =>
   await FavoriteMovies.create({ userID: id, ...movie });
@@ -18,8 +18,8 @@ const addFavoriteMovie = async (movie, id) =>
 const addWatchedMovie = async (movie, id) =>
   await WatchedMovies.create({ userID: id, ...movie });
 
-const removeFavoriteByID = async (movieID) =>
-  await FavoriteMovies.findByIdAndRemove({ _id: movieID });
+const removeFavoriteByID = async (userID, movieID) =>
+  await FavoriteMovies.findByIdAndRemove({ userId: userID, _id: movieID });
 
 const removeWatchedByID = async (movieID) =>
   await WatchedMovies.findByIdAndRemove({ _id: movieID });
@@ -27,8 +27,8 @@ const removeWatchedByID = async (movieID) =>
 module.exports = {
   findFavoritesByUserID,
   findWatchedByUserID,
-  findFavoriteByTitle,
-  findWatchedByTitle,
+  findFavoriteByMovieID,
+  findWatchedByMovieID,
   addFavoriteMovie,
   addWatchedMovie,
   removeFavoriteByID,
